Start refactoring SubsectorDrawer into three separate sub-forms
